requirements.txt
openai>=1.0.0
Pillow>=10.0.0
pillow-heif>=0.10.0  # HEIC/HEIF support for Pillow
python-dotenv>=1.0.0
tqdm>=4.65.0